+++ lib/teuton-get.rb
@@ -1,13 +1,12 @@
 # frozen_string_literal: true
 
-require "json/pure"
 require_relative "teuton-get/downloader"
 require_relative "teuton-get/repo/local_info"
 require_relative "teuton-get/repo/local_repo"
 require_relative "teuton-get/repo/repo_config"
 require_relative "teuton-get/repo/repo_data"
 require_relative "teuton-get/searcher"
-require_relative "teuton-get/writer/format"
+require_relative "teuton-get/show_info"
 
 module TeutonGet
   def self.create_info(testpath)
@@ -36,42 +35,23 @@ module TeutonGet
     RepoData.default.refresh
   end
 
-  def self.show_repo_list
-    RepoConfig.default.show_list
+  def self.show_repo_list(...)
+    RepoConfig.default.show_list(...)
   end
 
-  def self.show_info(test_id, options)
-    repo_data = RepoData.default
-    testinfo = repo_data.get_info(test_id)
-
-    if testinfo == {}
-      results = Searcher.default.get(test_id).results
-      if results.size == 1
-        test_id = results[0][:id]
-      else
-        puts "(#{results.size} results!)"
-        results.each { |i| puts "* #{i[:id]}" }
-        exit 1
-      end
-    end
-
-    testinfo = repo_data.get_info(test_id)
-    exit 1 if testinfo == {}
-
-    if options["format"] == "json"
-      puts testinfo.to_json
-    else
-      repo_data.show_testinfo(testinfo)
-    end
+  def self.show_info(...)
+    ShowInfo.new.call(...)
   end
 
   def self.search(filter, options)
     searcher = Searcher.default
-    results = searcher.get(filter)
+    search = searcher.get(filter)
+    exit 1 if search.results.size.zero?
+
     if options["format"] == "json"
-      results.show_json
+      search.show_json
     else
-      results.show_screen
+      search.show_screen
     end
   end
 end