+++ lib/teuton-get/repo/repo_config.rb
@@ -1,4 +1,5 @@
 require "fileutils"
+require "json/pure"
 require_relative "../settings"
 require_relative "../reader/inifile_reader"
 require_relative "../writer/format"
@@ -33,13 +34,24 @@ class RepoConfig
     create_ini_file
   end
 
-  def show_list
+  def show_list(options)
+    if options["format"] == "json"
+      show_json_list
+    else
+      show_default_list
+    end
+  end
+
+  private
+
+  def show_default_list
     rows = []
     rows << ["E", "NAME", "DESCRIPTION"]
     rows << :separator
 
     index = 0
     @data.each_pair do |key, value|
+      next unless value["type"] == "teutontest"
       enable = "\u{2714}"
       enable = " " unless value["enable"]
       description = value["description"] || "?"
@@ -52,7 +64,16 @@ class RepoConfig
     @dev.writeln "#{@reader.source}\n", color: :white
   end
 
-  private
+  def show_json_list
+    alist = []
+    @data.each_pair do |key, values|
+      next unless values["type"] == "teutontest"
+
+      values["reponame"] = key
+      alist << values
+    end
+    puts JSON.dump(alist)
+  end
 
   def create_dir
     dirpath = @config_dirpath