+++ CHANGELOG.md
@@ -7,8 +7,13 @@
 
 -- activar salida --format=json (search, info, repos)
 -- create-info sin argumentos asume el .
--- teutonget show nombre-incompleto pero único lo acepta
--- teutonget version
+
+## [0.3.0] - 2023-03-15
+
+-- New output option --format=json (teutonget search, teutonget show, teutonget repos)
+-- New output option --no-color
+-- teutonget show accept TESTID or FILTER
+-- teutonget version without color
 
 ## [0.2.7] - 2023-02-26